Refactored E2E tests to stabilize them
