+++ test/app-sync.e2e-spec.ts
@@ -3,38 +3,50 @@ import { INestMicroservice } from '@nestjs/common';
 import AppModule from './e2e/app/config.app.sync';
 import { TestConsumer } from './e2e/app/test.controller';
 import { messages } from './constants';
+import { Utils } from './utils';
 
 describe('AppModule Sync (e2e)', () => {
-  let app: INestMicroservice;
-  let controller: TestConsumer;
+    let app: INestMicroservice;
+    let controller: TestConsumer;
 
-  beforeAll(async () => {
-    const moduleFixture: TestingModule = await Test.createTestingModule({
-      imports: [AppModule],
-    }).compile();
+    beforeAll(async () => {
+        jest.setTimeout(10000);
 
-    app = moduleFixture.createNestMicroservice({});
-    app.enableShutdownHooks();
-    await app.listenAsync();
-    
-    controller = await app.resolve(TestConsumer);
-    controller.messages = [];
-  });
+        await Utils.schemaRegistrySetup();
 
-  afterAll(async() => {
-    await app.close();
-  });
+        const moduleFixture: TestingModule = await Test.createTestingModule({
+            imports: [AppModule]
+        }).compile();
 
-  it("should give kafka some time", done => {
-    setTimeout(done, 4000);
-  });
-
-  it('We can SEND and ACCEPT AVRO messages', async () => {
-    return controller.sendMessage({ messages }).then(() => {
-      expect(controller.messages.length).toBe(messages.length);
-
-      expect(controller.messages[0]).toEqual(messages[0].value);
-      expect(controller.messages[1]).toEqual(messages[1].value);
+        app = moduleFixture.createNestMicroservice({});
+        app.enableShutdownHooks();
+        await app.listenAsync();
     });
-  });
+
+    afterAll(async () => {
+        await app.close();
+    });
+
+    beforeEach(async () => {
+        await Utils.sleep(2000);
+
+        controller = await app.resolve(TestConsumer);
+        controller.messages = [];
+    });
+
+    it('We can SEND and ACCEPT AVRO messages', async () => {
+        await Utils.sleep(2000);
+
+        await controller.sendMessage({ messages });
+
+        let count = 0;
+        while (controller.messages.length < messages.length && count < 4) {
+            await Utils.sleep(1000);
+            count++;
+        }
+
+        expect(controller.messages.length).toBe(messages.length);
+        expect(controller.messages.find((x) => x.id == messages[0].value.id)).toBeDefined();
+        expect(controller.messages.find((x) => x.id == messages[1].value.id)).toBeDefined();
+    });
 });
